llm-translation-test/requirements.txt
accelerate==1.13.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
anyio==4.12.1
attrs==26.1.0
certifi==2026.2.25
charset-normalizer==3.4.6
colorama==0.4.6
cuda-bindings==12.9.4
cuda-pathfinder==1.4.3
datasets==4.8.3
dill==0.4.1
diskcache==5.6.3
entmax==1.3
filelock==3.25.2
frozenlist==1.8.0
fsspec==2026.2.0
h11==0.16.0
hf-xet==1.4.2
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.36.2
idna==3.11
jinja2==3.1.6
jsonargparse==3.13.1
lightning-utilities==0.15.3
llama-cpp-python==0.3.16
lxml==6.0.2
markupsafe==3.0.3
mpmath==1.3.0
multidict==6.7.1
multiprocess==0.70.19
networkx==3.6.1
numpy==1.26.4
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.4.5
nvidia-nvtx-cu12==12.8.90
packaging==26.0
pandas==2.3.3
portalocker==3.2.0
propcache==0.4.1
protobuf==4.25.8
psutil==7.2.2
pyarrow==23.0.1
python-dateutil==2.9.0.post0
pytorch-lightning==2.6.1
pytz==2026.1.post1
pyyaml==6.0.3
regex==2026.2.28
requests==2.32.5
sacrebleu==2.6.0
safetensors==0.7.0
scipy==1.17.1
sentencepiece==0.2.1
setuptools==82.0.1
six==1.17.0
sympy==1.14.0
tabulate==0.10.0
tokenizers==0.22.2
torch==2.10.0
torchmetrics==0.10.3
tqdm==4.67.3
transformers==4.57.6
triton==3.6.0
typing-extensions==4.15.0
tzdata==2025.3
unbabel-comet==2.2.7
urllib3==2.6.3
xxhash==3.6.0
yarl==1.23.0
